requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
pandas<2.0.0
numpy<2.0.0
scikit-learn>=1.2.2
scikit-surprise>=1.1.3
joblib>=1.2.0
python-multipart>=0.0.6
